Update to 1.9
