+++ .gitignore
@@ -24,3 +24,4 @@ nb-configuration.xml
 *\.tex~
 *\.yml~
 *.synctex.gz
+/target/
