+++ src/main/java/org/communitybridge/achievement/AchievementAvatar.java
