+++ src/main/java/org/communitybridge/grouprules/GroupRuleDirection.java
