+++ src/main/java/org/communitybridge/main/BukkitWrapper.java
@@ -1,6 +1,8 @@
 package org.communitybridge.main;
 
+import java.util.Collection;
 import java.util.UUID;
+
 import org.bukkit.BanList;
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
@@ -15,6 +17,7 @@ public class BukkitWrapper
 		return Bukkit.getOfflinePlayer(uuid);
 	}
 
+	@SuppressWarnings("deprecation")
 	public OfflinePlayer getOfflinePlayer(String name)
 	{
 		return Bukkit.getOfflinePlayer(name);
@@ -40,7 +43,7 @@ public class BukkitWrapper
 		return Bukkit.getPluginManager();
 	}
 
-	public Player[] getOnlinePlayers()
+	public Collection<? extends Player> getOnlinePlayers()
 	{
 		return Bukkit.getOnlinePlayers();
 	}
