+++ src/main/java/org/communitybridge/main/CBMetrics.java
@@ -347,7 +347,7 @@ public class CBMetrics {
         boolean onlineMode = Bukkit.getServer().getOnlineMode(); // TRUE if online mode is enabled
         String pluginVersion = description.getVersion();
         String serverVersion = Bukkit.getVersion();
-        int playersOnline = Bukkit.getServer().getOnlinePlayers().length;
+        int playersOnline = Bukkit.getServer().getOnlinePlayers().size();
 
         // END server software specific section -- all code below does not use any code outside of this class / Java
 
@@ -505,7 +505,7 @@ public class CBMetrics {
      * @param input
      * @return
      */
-		@SuppressWarnings("PMD.EmptyCatchBlock")
+
     public static byte[] gzip(String input) {
         ByteArrayOutputStream baos = new ByteArrayOutputStream();
         GZIPOutputStream gzos = null;
