+++ src/main/java/org/communitybridge/synchronization/Synchronizer.java
