+++ src/main/java/org/communitybridge/synchronization/ban/BanState.java
