+++ src/main/java/org/communitybridge/utility/MinecraftUtilities.java
