+++ src/main/java/org/communitybridge/utility/StringUtilities.java
